+++ packages/visual-grid-client/src/sdk/checkWindow.js
@@ -87,7 +87,7 @@ function makeCheckWindow({
       floating,
     })
 
-    const resourcesPromises = snapshots.map(async snapshot => {
+    const resourcesPromises = snapshots.map(async (snapshot, index) => {
       const {rGridDom: dom, allResources: resources} = await createRGridDOMAndGetResourceMapping({
         resourceUrls: snapshot.resourceUrls,
         resourceContents: snapshot.resourceContents,
@@ -96,6 +96,7 @@ function makeCheckWindow({
         userAgent,
         referer: url,
         proxySettings: wrappers[0].getProxy(),
+        browserName: browsers[index].name,
       })
       await putResources([dom, ...Object.values(resources)])
       return {dom, resources: Object.values(resources)}