+++ packages/visual-grid-client/src/sdk/createRGridDOMAndGetResourceMapping.js
@@ -11,6 +11,7 @@ function makeCreateRGridDOMAndGetResourceMapping({getAllResources}) {
     userAgent,
     referer,
     proxySettings,
+    browserName,
   }) {
     const resources = await getAllResources({
       resourceUrls,
@@ -18,6 +19,7 @@ function makeCreateRGridDOMAndGetResourceMapping({getAllResources}) {
       userAgent,
       referer,
       proxySettings,
+      browserName,
     })
     const allResources = Object.assign({}, resources)